Kraken private DTO Object comments
